+++ tools/corda-runtime-gradle-plugin/src/main/kotlin/net/corda/gradle/plugin/ProjectUtils.kt
@@ -6,6 +6,8 @@ import kong.unirest.HttpResponse
 import kong.unirest.JsonNode
 import kong.unirest.Unirest
 import net.corda.gradle.plugin.configuration.ProjectContext
+import net.corda.gradle.plugin.dtos.CpiMetadataDTO
+import net.corda.gradle.plugin.dtos.GetCPIsResponseDTO
 import net.corda.gradle.plugin.dtos.VirtualNodeInfoDTO
 import net.corda.gradle.plugin.dtos.VirtualNodesDTO
 import net.corda.gradle.plugin.exception.CordaRuntimeGradlePluginException
@@ -90,6 +92,27 @@ fun getExistingNodes(pc: ProjectContext) : List<VirtualNodeInfoDTO> {
 
     return try {
         mapper.readValue(response.body.toString(), VirtualNodesDTO::class.java).virtualNodes!!
+    } catch (e: Exception) {
+        throw CordaRuntimeGradlePluginException("Failed to get Existing vNodes with exception: $e")
+    }
+}
+
+fun getUploadedCpis(pc: ProjectContext): List<CpiMetadataDTO> {
+
+    Unirest.config().verifySsl(false)
+    val mapper = ObjectMapper()
+    mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+
+    val response: HttpResponse<JsonNode> = Unirest.get(pc.cordaClusterURL + "/api/v1/cpi")
+        .basicAuth(pc.cordaRpcUser, pc.cordaRpcPassword)
+        .asJson()
+
+    if (response.status != HttpURLConnection.HTTP_OK) {
+        throw CordaRuntimeGradlePluginException("Failed to get Existing Cpis, response status: " + response.status)
+    }
+
+    return try {
+        mapper.readValue(response.body.toString(), GetCPIsResponseDTO::class.java).cpis!!
     } catch (e: Exception) {
         throw CordaRuntimeGradlePluginException("Failed to get Existing vNodes with exception: $e")
     }
