--- tools/corda-runtime-gradle-plugin/src/main/kotlin/net/corda/gradle/plugin/dtos/GetCPIsResponseDTO.kt
@@ -1,5 +1,0 @@
-package net.corda.gradle.plugin.dtos
-
-class GetCPIsResponseDTO {
-    var cpis: List<CpiMetadataDTO>? = null
-}
