--- tools/corda-runtime-gradle-plugin/src/main/kotlin/net/corda/gradle/plugin/dtos/VirtualNodesDTO.kt
@@ -1,5 +1,0 @@
-package net.corda.gradle.plugin.dtos
-
-class VirtualNodesDTO {
-    var virtualNodes: List<VirtualNodeInfoDTO>? = null
-}
