+++ tools/corda-runtime-gradle-plugin/src/test/kotlin/net/corda/gradle/plugin/CordaGradlePluginTest.kt
@@ -1,10 +1,8 @@
 package net.corda.gradle.plugin
 
-//import net.corda.craft5.annotations.TestSuite
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 
-//@TestSuite
 class CordaGradlePluginTest : FunctionalBaseTest() {
 
     @Test
